requirements.txt
fastapi
uvicorn
mcp
python-dotenv
requests
